trying to split syz_emit_ethernet & modifed seeds now the syz_emit_ethernet pseudo syscall packs all ethernet packet types into one single syscall. This is not efficient if we want to fuzz only one specific protocol.Trying to split by defining syscall variants, start from TCP, working on
the TCP handshake seeds is not working as expected after carefully
checking the coverage info. The TCP-SynAck packet was sent to the tun
device, but syz_extract_tcp_res() failed to extract the ack number. By
checking the coverage info of TUN device, the syscall got stuck
in the if(noblock) check in tun_put_user() and returned too early. This
seems to be resulted from opening the tun device file with O_NONBLOCK
flag in initialize_tun()

but opening the device without this flag caused Syzkaller to fail to get
coverage info, no idea why